Published latest distribution.
There is still an issue with the Fla files and fonts when these are
published on Mac: for some reason the fonts embedded in the project
asset SWC are not visible in the compiled demo. The Windows-published
SWC works nicely, just not the Mac one, even though the required fonts
seem to be installed correctly and valid.