Added support to build process for building Desktop builds with non-packed resource files.
Made Android and iOS build use non-packed resource files as these aren't
really necessary for mobile builds (on mobiles it is common to use
non-packed asset files and the resource unpacking decreases load
performance).

Organized the build task names to be in better order and easier to
distinct.

Made changes to resource system to account for non-packed resource files
in AIR builds.